Added support for MD5 dataheader cache.
The question here is if MD5 is really the way to go since we do not need to check data integrity but only signify that dataheader changed. Since sender _always_ knows if the data header has changed the MD5 checksum could also be replaced by a simple counter that is incresed whenever the dataheader is changed. This would simplify both the sender and receiver (both lose md5 dependency) and functionally nothing would be lost.